+++ lib/src/wpp/wpp_events.dart
@@ -127,8 +127,4 @@ class WppEvents {
       }
     });
   }
-
-  Future<void> _removeEventListeners() async {
-    // TODO : remove all Event Listeners
-  }
 }